modify

+++ trabalhofinalanalipe/database/migrations/2023_09_11_164314_create_postagem_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string("titulo");
             $table->string("conteudo");
             $table->string("imagem")->nullable();
-            $table->int("curtidas");
+            $table->bigInteger("curtidas");
             $table->bigInteger("user_id")->unsigned();
             $table->timestamps();
